+++ python/pathway/internals/column_namespace.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from pathway.internals.join import Joinable
+    from pathway.internals.joins import Joinable
 
 
 class ColumnNamespace: