+++ python/pathway/stdlib/temporal/_asof_now_join.py
@@ -14,7 +14,7 @@ from pathway.internals.desugaring import (
     TableSubstitutionDesugaring,
     desugar,
 )
-from pathway.internals.join import validate_join_condition
+from pathway.internals.joins import validate_join_condition
 from pathway.internals.runtime_type_check import check_arg_types
 from pathway.internals.thisclass import ThisMetaclass
 from pathway.internals.trace import trace_user_frame